+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from "./codegen";
+export * from "./http";
 export * from "./introspect";
 export * from "./read";
 export * from "./server";