+++ src/server/buildExecutableSchema.ts
@@ -1,6 +1,6 @@
 import { GraphQLFieldResolver, GraphQLSchema, isSchema } from "graphql";
 import { join } from "path";
-import { GraphQLReader } from "../read";
+import { GraphQLReader } from "../read/GraphQLReader";
 import {
   GraphQLSchemaManager,
   SubscriptionResolver,