+++ src/server/index.ts
@@ -1,4 +1,3 @@
 export * from "./GraphQLExecutionArgsParser";
 export * from "./GraphQLSchemaManager";
-export * from "./GraphQLServer";
 export * from "./buildExecutableSchema";