+++ src/util/index.ts
@@ -3,6 +3,6 @@ export * from "./AsyncPushIterator";
 export * from "./cleanOperations";
 export * from "./DeferredPromise";
 export * from "./directives";
-export * from "./misc";
+export * from "./isAsyncIterator";
 export * from "./watchRecursive";
 export * from "./writeTypeScript";