+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 
 ## vNEXT
 
-- Allow creating a context per web socket subscription (breaking change)
-- Refactor auto-reconnect API in client web sockets (breaking change)
-- Improve auto-reconnection in client web sockets
+- Refactor web socket package
+  - Allow creating a context per web socket subscription (breaking change)
+  - Refactor auto-reconnect API in client web sockets (breaking change)
+  - Improve auto-reconnection in client web sockets
+- Refactor package structure
 - Fix UMD build
 - Improved types for resolver maps
 - Add some doc-block comments
+- Enable subscriptions in example GraphiQL
 
 ## v0.1.2